Remove annotate fasta, fix test paths

+++ CAMPER_DRAMKit/setup.py
@@ -8,17 +8,18 @@ __version__ = version
 
 here = os.path.abspath(os.path.dirname(__file__))
 data_folder = os.path.join(here, 'camper_dramkit', 'data')
-os.mkdir(data_folder)
-os.rename(os.path.join(here, '..', 'CAMPER_blast.faa'),
-          os.path.join(data_folder, 'CAMPER_blast.faa'))
-os.rename(os.path.join(here, '..', 'CAMPER_blast_scores.tsv'),
-          os.path.join(data_folder, 'CAMPER_blast_scores.tsv'))
-os.rename(os.path.join(here, '..', 'CAMPER.hmm'),
-          os.path.join(data_folder, 'CAMPER.hmm'))
-os.rename(os.path.join(here, '..', 'CAMPER_distillate.tsv'),
-          os.path.join(data_folder, 'CAMPER_distillate.tsv'))
-os.rename(os.path.join(here, '..', 'CAMPER_hmm_scores.tsv'),
-          os.path.join(data_folder, 'CAMPER_hmm_scores.tsv'))
+if not os.path.exists(data_folder):
+    os.mkdir(data_folder)
+    os.rename(os.path.join(here, '..', 'CAMPER_blast.faa'),
+              os.path.join(data_folder, 'CAMPER_blast.faa'))
+    os.rename(os.path.join(here, '..', 'CAMPER_blast_scores.tsv'),
+              os.path.join(data_folder, 'CAMPER_blast_scores.tsv'))
+    os.rename(os.path.join(here, '..', 'CAMPER.hmm'),
+              os.path.join(data_folder, 'CAMPER.hmm'))
+    os.rename(os.path.join(here, '..', 'CAMPER_distillate.tsv'),
+              os.path.join(data_folder, 'CAMPER_distillate.tsv'))
+    os.rename(os.path.join(here, '..', 'CAMPER_hmm_scores.tsv'),
+              os.path.join(data_folder, 'CAMPER_hmm_scores.tsv'))
 
 with open(os.path.join(here, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
@@ -33,7 +34,7 @@ setup(
     package_dir={'camper_dramkit': 'camper_dramkit'},
     package_data={'camper_dramkit': ['data/*']},
     python_requires='>=3.8',
-    install_requires=['scikit-bio', 'pandas', 'altair', 'sqlalchemy', 'networkx', 'openpyxl', 'numpy', 'click'],
+    install_requires=['pandas', 'altair', 'sqlalchemy', 'networkx', 'openpyxl', 'numpy', 'click'],
     entry_points={
         'console_scripts': [
             'camper_distill = camper_dramkit.camper_distill:summarize_genomes',
